TechFiesta/requirement.txt
fastapi
uvicorn[standard]
python-multipart
pypdf2
python-docx
sentence-transformers
scikit-learn
nltk
radon
tavily-python==0.3.8
requests


